+++ test/support/lib/drab_test_app/endpoint.ex
@@ -32,7 +32,7 @@ defmodule DrabTestApp.Endpoint do
     Plug.Parsers,
     parsers: [:urlencoded, :multipart, :json],
     pass: ["*/*"],
-    json_decoder: Poison
+    json_decoder: Jason
   )
 
   plug(Plug.MethodOverride)